% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{botrytis.lr.applied.riesling}
\alias{botrytis.lr.applied.riesling}
\title{Linear regression applied for calculating the CDD_7;18;24 reaching 5\% disease severity (Botrytis cinerea) for Riesling}
\usage{
botrytis.lr.applied.riesling(
  stat.tmean01,
  stat.tmean02,
  stat.tmean03,
  stat.rain01,
  stat.rain02,
  coef
)
}
\arguments{
\item{stat.tmean01}{numeric, vector with two columns: 1) year, and 2) first statistic for mean temperature.}

\item{stat.tmean02}{numeric, vector with two columns: 1) year, and 2) second statistic for mean temperature.}

\item{stat.tmean03}{numeric, vector with two columns: 1) year, and 2) third statistic for mean temperature.}

\item{stat.rain01}{numeric, vector with two columns: 1) year, and 2) first statistic for precipitation.}

\item{stat.rain02}{numeric, vector with two columns: 1) year, and 2) second statistic for precipitation.}

\item{coef}{the predefined linear regression coefficients,  1 observation of 6 variables, being the first
the intercept, and the remainder variables the coefficients for each statistic matching the definition order.}
}
\value{
data.frame, 7 obs. of 2 variables (year and CDD for 7, 18 and 24 thresholds reaching 5\% disease severity,
Botrytis cinerea, for Riesling).
}
\description{
Linear regression applied for calculating the CDD_7;18;24 reaching 5\% disease severity (Botrytis cinerea) for Riesling
}
