% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenology.R
\name{cdd.lThresh.phenology}
\alias{cdd.lThresh.phenology}
\title{Cumulative degree days (cdd) by the single (lower) threshold algorithm for phenology}
\usage{
cdd.lThresh.phenology(cdd.lt, chs.mean)
}
\arguments{
\item{cdd.lt}{list, cumulative degree days (in Celsius degrees) for vine growth in xts format per year
as provided by "cdd.1Treshold" function.}

\item{chs.mean}{numeric, mean cumulative heat sum for bud break (in Celsius degrees).}
}
\value{
the cumulative degree days (in Celsius degrees) for vine growth plus an additional column with
the cumulative degree days (in Celsius degrees) for phenology.
}
\description{
Implementation to compute the cumulative degree days (cdd) by the single (lower) threshold algorithm by
Molitor et al. (2014) for phenology.
}
\references{
Daniel Molitor, Jürgen Junk, Danièle Evers, Lucien Hoffmann, and Marco Beyer (2014).
A high-resolution cumulative degree day-based model to simulate phenological development of grapevine.
Am. J. Enol. Vitic., (65:1):72–80.
}
