% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenology.R
\name{cdd.luThresh.phenology}
\alias{cdd.luThresh.phenology}
\title{Cumulative degree days (CDD) by the double (lower and upper) threshold temperatures for phenology}
\usage{
cdd.luThresh.phenology(cdd.lut, chs.mean)
}
\arguments{
\item{cdd.lut}{list, cumulative degree days (in Celsius degrees) for vine growth in xts format per year
as provided by the double threshold temperature ("cdd.double.threshold" function).}

\item{chs.mean}{numeric, mean cumulative heat sum for bud break (in Celsius degrees).}
}
\value{
the cumulative degree days (in Celsius degrees) for vine growth plus an additional column with
the cumulative degree days (in Celsius degrees) for phenology.
}
\description{
Implementation to compute the cumulative degree days by the double (lower and upper) threshold temperatures by
Molitor et al. (2014) for phenology.
}
\references{
Daniel Molitor, Jürgen Junk, Danièle Evers, Lucien Hoffmann, and Marco Beyer (2014).
A high-resolution cumulative degree day-based model to simulate phenological development of grapevine.
Am. J. Enol. Vitic., (65:1):72–80.
}
