% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudBurst.R
\name{cdd.simple.budbreak}
\alias{cdd.simple.budbreak}
\title{Cumulative degree days (CDD) by the simple algorithm}
\usage{
cdd.simple.budbreak(data, t.mean.col, start.month, start.day)
}
\arguments{
\item{data}{input data in xts format.}

\item{t.mean.col}{numeric, column position in data for the daily mean air temperature vector in Celsius degrees.}

\item{start.day}{}
}
\value{
list per year for the input data plus an additional column with the cumulative degree days
(in Celsius degrees) for vine growth. The output for each year is a "xts" time series object.
}
\description{
Implementation to compute the cumulative degree days by the simple algorithm by Nendel (2010).
}
\references{
Nendel, Class (2010). Grapevine bud break prediction for cool winter climates.
Int. J. Biometeorol., 54, 231–241.
}
