% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudBurst.R
\name{cdd.single.triangle.budbreak}
\alias{cdd.single.triangle.budbreak}
\title{Cumulative degree days (CDD) by the single triangle algorithm for bud break}
\usage{
cdd.single.triangle.budbreak(cdd, start.month, start.day)
}
\arguments{
\item{cdd}{cumulative degree days (in Celsius degrees) for vine growth in xts format as provided by
"cdd.single.triangle" function.}

\item{start.month}{numeric, calculated optimum starting month of year.}

\item{start.day}{numeric, calculated optimum starting day of start.month.}
}
\value{
the cumulative degree days (in Celsius degrees) for vine growth plus an additional column with
the cumulative degree days (in Celsius degrees) for bud break.
}
\description{
Implementation to compute the cumulative degree days by the single triangle algorithm by Nendel (2010)
for bud break.
}
\references{
Nendel, Class (2010). Grapevine bud break prediction for cool winter climates.
Int. J. Biometeorol., 54, 231–241.
}
