% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenology.R
\name{compare.stage}
\alias{compare.stage}
\title{Compare by growth stage from phenology output}
\usage{
compare.stage(ref.data, phen, growth.stage)
}
\arguments{
\item{ref.data}{data.frame, reference dataset to define the observations for the phenological stage to compare
e.g. "data_remich_bbch09" or "data_remich_bbch81" datasets.}

\item{phen}{list per year, with each list containing a data.frame with the phenological stages
for vine growth as the output from the "phenology.stages" function.}

\item{growth.stage}{numeric, one of the growth stages to compute the summary.}
}
\value{
data.frame, with each row containing a year with the comparison of the phenological stage computed
and observed. The last colunm indicates the "Difference" between observed and computed day of year (DOY).
}
\description{
Implementation to compare observations versus computations for growth stage from phenology output.
}
