% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{data_boundary_grevenmacher}
\alias{data_boundary_grevenmacher}
\title{A 'sp' object for the Grevenmacher admininistrative boundaries}
\format{Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots \cr
..@ data       :'data.frame':	23 obs. of  4 variables: \cr
 .. ..$ COMMUNE : Factor w/ 23 levels "Beaufort","Bech",..: 14 16 1 18 9 12 8 10 4 2 ... \cr
 .. ..$ CANTON  : Factor w/ 3 levels "Echternach","Grevenmacher",..: 2 3 1 1 1 2 3 2 2 1 ... \cr
 .. ..$ DISTRICT: Factor w/ 1 level "Grevenmacher": 1 1 1 1 1 1 1 1 1 1 ... \cr
 .. ..$ LAU2    : Factor w/ 23 levels "1001","1002",..: 13 19 1 6 5 12 17 10 8 2 ... \cr
 ..@ polygons   :List of 23 \cr
 .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots \cr
 .. .. .. ..@ Polygons :List of 1 \cr
 .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots \cr
 .. .. .. .. .. .. ..@ labpt  : num [1:2] 97715 87463 \cr
 .. .. .. .. .. .. ..@ area   : num 27853866 \cr
 .. .. .. .. .. .. ..@ hole   : logi FALSE \cr
 .. .. .. .. .. .. ..@ ringDir: int 1 \cr
 #'  .. .. .. .. .. .. ..@ coords : num [1:1611, 1:2] 96023 96027 96037 96070 96078 ... \cr
 .. .. .. ..@ plotOrder: int 1 \cr
 .. .. .. ..@ labpt    : num [1:2] 97715 87463 \cr
 .. .. .. ..@ ID       : chr "0" \cr
 .. .. .. ..@ area     : num 27853866 \cr
 .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots \cr
 .. .. .. ..@ Polygons :List of 1 \cr
 .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots \cr
 .. .. .. .. .. .. ..@ labpt  : num [1:2] 88254 64603 \cr
 .. .. .. .. .. .. ..@ area   : num 13591512 \cr
 .. .. .. .. .. .. ..@ hole   : logi FALSE \cr
 .. .. .. .. .. .. ..@ ringDir: int 1 \cr
 .. .. .. .. .. .. ..@ coords : num [1:1012, 1:2] 89679 89677 89674 89673 89673 ... \cr
 .. .. .. ..@ plotOrder: int 1 \cr
 .. .. .. ..@ labpt    : num [1:2] 88254 64603 \cr
 .. .. .. ..@ ID       : chr "1" \cr
 .. .. .. ..@ area     : num 13591512 \cr
 .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots \cr
 .. .. .. ..@ Polygons :List of 1 \cr
 .. .. .. ..[list output truncated]}
\source{
\url{https://data.public.lu/fr/datasets/limites-administratives-du-grand-duche-de-luxembourg/}
}
\usage{
data(data_boundary_grevenmacher)
}
\description{
A dataset containing a 'sp' object for the admininistrative boundaries of the District of Grevenmacher
in the Grand-Duchy of Luxembourg. It comprises the Commune and Canton levels as provided by the Luxembourgish Data
Platform. The data is provided as a SpatialPolygonsDataFrame from "sp" package.
}
\details{
The spatial representation for the data corresponds to a SpatialPolygonsDataFrame with
coordinate reference system (CRS) EPSG:2169 - Luxembourg 1930 / Gauss - Projected.
}
\examples{
data(data_boundary_grevenmacher)
str(data_boundary_grevenmacher)

#' plot stfdf object
plot(data_boundary_grevenmacher)

}
\keyword{data}
