% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{data_lm_yield_rivaner}
\alias{data_lm_yield_rivaner}
\title{Yield model for Rivaner cultivar}
\format{
List of 12  \cr
}
\usage{
data(data_lm_yield_rivaner)
}
\description{
A sample dataset containing the Linear regression model calculated by
Daniel Molitor and Markus Keller (2016) for predicting the yield
for Rivaner cultivar.
}
\references{
and Markus Keller (2016). Yield of Müller-Thurgau and Riesling grapevines
is altered by meteorological conditions in the current and previous growing seasons
OENO One, 2016, 50, 4, 245 - 258.
}
\keyword{data}
