% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{data_vineyards2018}
\alias{data_vineyards2018}
\title{A 'sp' object for the vineyards along the Mosel in Luxemburg in 2018}
\format{
Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots \cr
..@ data       :'data.frame':	4966 obs. of  7 variables: \cr
  .. ..$ Weinbergsn: Factor w/ 4950 levels "1","10005","10007",..: 1359 4035 3194 2950 2952 1083 2849 2871 2874 2951 ... \cr
  .. ..$ CODE_ELEM : Factor w/ 1 level "V": 1 1 1 1 1 1 1 1 1 1 ... \cr
  .. ..$ CODE_COM  : Factor w/ 14 levels "?","022","023",..: 12 4 14 6 12 12 14 14 14 14 ... \cr
  .. ..$ CODE_SECT : Factor w/ 8 levels "A","b","B","c",..: 1 7 8 7 3 3 7 7 7 7 ... \cr
  .. ..$ Shape_Leng: num [1:4966] 164 247 231 311 282 ... \cr
  .. ..$ Shape_Le_1: num [1:4966] 164 247 231 311 282 ... \cr
  .. ..$ Shape_Area: num [1:4966] 504 3773 3000 6016 2686 ... \cr
  ..@ polygons   :List of 4966 \cr
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots \cr
  .. .. .. ..@ Polygons :List of 1 \cr
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots \cr
  .. .. .. .. .. .. ..@ labpt  : num [1:2] 93823 71477 \cr
  .. .. .. .. .. .. ..@ area   : num 504 \cr
  .. .. .. .. .. .. ..@ hole   : logi FALSE \cr
  .. .. .. .. .. .. ..@ ringDir: int 1 \cr
  .. .. .. .. .. .. ..@ coords : num [1:6, 1:2] 93793 93797 93853 93851 93849 ... \cr
  .. .. .. ..@ plotOrder: int 1 \cr
  .. .. .. ..@ labpt    : num [1:2] 93823 71477 \cr
  .. .. .. ..@ ID       : chr "0" \cr
  .. .. .. ..@ area     : num 504 \cr
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots \cr
  .. .. .. ..@ Polygons :List of 1 \cr
  .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots \cr
  .. .. .. .. .. .. ..@ labpt  : num [1:2] 96817 77073 \cr
  .. .. .. .. .. .. ..@ area   : num 3773 \cr
  .. .. .. .. .. .. ..@ hole   : logi FALSE \cr
  .. .. .. .. .. .. ..@ ringDir: int 1 \cr
  .. .. .. .. .. .. ..@ coords : num [1:8, 1:2] 96857 96848 96839 96827 96777 ... \cr
  .. .. .. ..@ plotOrder: int 1 \cr
  .. .. .. ..@ labpt    : num [1:2] 96817 77073 \cr
  .. .. .. ..@ ID       : chr "1" \cr
  .. .. .. ..@ area     : num 3773 \cr
  .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots \cr
  .. .. .. ..@ Polygons :List of 1 \cr
  .. .. .. ..[list output truncated]
}
\source{
\url{https://data.public.lu/fr/datasets/vineyards/}
}
\usage{
data(data_vineyards2018)
}
\description{
A dataset containing a 'sp' object for the vineyards along the Mosel in the Grand-Duchy of Luxembourg in 2018.
The data is provided by the Luxembourgish Data Platform, and provided as a SpatialPolygonsDataFrame from "sp" package.
}
\details{
The spatial representation for the data corresponds to a SpatialPolygonsDataFrame with
coordinate reference system (CRS) EPSG:2169 - Luxembourg 1930 / Gauss - Projected. \cr
}
\examples{
data(data_vineyards2018)
str(data_vineyards2018, list.len = 15)

#' plot sp object
library(sp)
spplot(data_vineyards2018, zcol="Weinbergsn", col = "Brown", scales = list(draw = TRUE),
       colorkey = FALSE)

}
\keyword{data}
