% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{data_water_surface}
\alias{data_water_surface}
\title{A 'sp' object for the Grevenmacher water surfaces}
\format{Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots \cr
..@ data       :'data.frame':	1373 obs. of  5 variables: \cr
 .. ..$ cat     : int [1:1373] 1 2 3 4 5 6 7 8 9 10 ... \cr
 .. ..$ id      : int [1:1373] 1 1 1 1 1 1 1 1 1 1 ... \cr
 .. ..$ ID_2    : Factor w/ 1353 levels "2?443?282","2?443?944",..: 2 4 9 16 19 35 39 45 51 56 ... \cr
 .. ..$ NATURE  : int [1:1373] 3 3 3 3 3 3 3 3 3 3 ... \cr
 .. ..$ TOPONYME: Factor w/ 4 levels "Bassin d'?puration",..: NA NA NA NA NA NA NA NA NA NA ... \cr
 ..@ polygons   :List of 1373 \cr
 .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots \cr
 .. .. .. ..@ Polygons :List of 1 \cr
 .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots \cr
 .. .. .. .. .. .. ..@ labpt  : num [1:2] 93462 66333 \cr
 .. .. .. .. .. .. ..@ area   : num 145 \cr
 .. .. .. .. .. .. ..@ hole   : logi FALSE \cr
 .. .. .. .. .. .. ..@ ringDir: int 1 \cr
 .. .. .. .. .. .. ..@ ringDir: int 1 \cr
 .. .. .. .. .. .. ..@ coords : num [1:11, 1:2] 93468 93468 93466 93464 93462 ... \cr
 .. .. .. ..@ plotOrder: int 1 \cr
 .. .. .. ..@ labpt    : num [1:2] 93462 66333 \cr
 .. .. .. ..@ ID       : chr "0" \cr
 .. .. .. ..@ area     : num 145 \cr
 .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots \cr
 .. .. .. ..@ Polygons :List of 1 \cr
 .. .. .. ..[list output truncated]}
\source{
\url{  https://data.public.lu/en/datasets/bd-l-tc-surface-deau/#_}
}
\usage{
data(data_water_surface)
}
\description{
A dataset containing a 'sp' object for the water surfaces in the District of Grevenmacher
in the Grand-Duchy of Luxembourg. The dataset contains streches from the Moselle (Mosel) and the S{\^u}re rivers.
The data is provided by the Luxembourgish Data Platform, and provided as a SpatialPolygonsDataFrame from "sp" package.
}
\details{
The spatial representation for the data corresponds to a SpatialPolygonsDataFrame with
coordinate reference system (CRS) EPSG:2169 - Luxembourg 1930 / Gauss - Projected.
}
\examples{
data(data_water_surface)
str(data_water_surface, list.len = 15)

#' plot stfdf object
library(sp)
spplot(data_water_surface, zcol="id", scales = list(draw = TRUE),
       colorkey = FALSE)

}
\keyword{data}
