% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenology.R
\name{DD.LUH.cumulative}
\alias{DD.LUH.cumulative}
\title{Compute cumulative degree days by the lower, upper and heat temperature thresholds}
\usage{
DD.LUH.cumulative(dd, start.cdd, cdd)
}
\arguments{
\item{dd}{vector with the degree days (in Celsius degrees) for vine growth in xts format as provided by
"DD.LUH" function.}

\item{start.cdd}{numeric, mean cumulative heat sum (in Celsius Degrees) for bud break.}

\item{cdd}{vector with the cumulative degree days (in Celsius degrees) for bud break.}
}
\value{
the cumulative degree days (in Celsius degrees) for vine growth plus an additional column with
the comulative degree days (in Celsius degrees) for bud break.
}
\description{
Implementation to compute cumulative degree days by the lower, upper and heat temperature threshold
by Molitor et al., (2014).
}
\references{
Nendel, Class (2010). Grapevine bud break prediction for cool winter climates.
Int. J. Biometeorol., 54, 231–241.

Daniel Molitor, Jürgen Junk, Danièle Evers, Lucien Hoffmann, and Marco Beyer (2014).
A high-resolution cumulative degree day-based model to simulate phenological development of grapevine.
Am. J. Enol. Vitic., (65:1):72–80.
}
