% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenology.R
\name{DD.single.cumulative}
\alias{DD.single.cumulative}
\title{Compute the cumulative degree days by the single temperature threshold}
\usage{
DD.single.cumulative(data, t.mean.col, a)
}
\arguments{
\item{data}{input data in xts format.}

\item{t.mean.col}{numeric, column position in data for the daily mean air temperature vector in Celsius degrees.}

\item{a}{numeric, threshold temperature (in Celsius degrees) for vine growth.}
}
\value{
list per year for the input data plus an additional column with the cumulative degree days
(in Celsius degrees) for vine growth. The output for each year is a "xts" time series object.
}
\description{
Implementation to compute the cumulative degree days by the single temperature threshold by Molitor et al., (2014).
}
\references{
Daniel Molitor, Jürgen Junk, Danièle Evers, Lucien Hoffmann, and Marco Beyer (2014).
A high-resolution cumulative degree day-based model to simulate phenological development of grapevine.
Am. J. Enol. Vitic., (65:1):72–80.
}
