% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudBurst.R
\name{DD.single.threshold}
\alias{DD.single.threshold}
\title{Compute the degree-days by the single temperature threshold}
\usage{
DD.single.threshold(t.mean, a)
}
\arguments{
\item{t.mean}{daily mean air temperature vector in Celsius degrees.}

\item{a}{numeric, threshold temperature (in Celsius degrees) for vine growth.}
}
\value{
a vector with the degree-days (in Celsius degrees) for vine growth.
}
\description{
Implementation to compute the degree-days by the single temperature threshold by Molitor et al., (2014).
}
\references{
Daniel Molitor, Jürgen Junk, Danièle Evers, Lucien Hoffmann, and Marco Beyer.
A high-resolution cumulative degree day-based model to simulate phenological development of grapevine.
Am. J. Enol. Vitic., (65:1):72–80, 2014.
}
