% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenology.R
\name{DD.single.treshold.phenology}
\alias{DD.single.treshold.phenology}
\title{Cumulative degree days (CDD) by the single threshold algorithm for phenology}
\usage{
DD.single.treshold.phenology(cdd.stt, chs.mean)
}
\arguments{
\item{cdd.stt}{cumulative degree days (in Celsius degrees) for vine growth in xts format as provided by
"DD.single.treshold.cumulative" function.}

\item{chs.mean}{numeric, mean cumulative heat sum for bud break (in Celsius degrees).}
}
\value{
the cumulative degree days (in Celsius degrees) for vine growth plus an additional column with
the comulative degree days (in Celsius degrees) for phenology.
}
\description{
Implementation to compute the cumulative degree days by the single threshold algorithm by
Molitor et al. (2010) for phenology.
}
\references{
Daniel Molitor, Jürgen Junk, Danièle Evers, Lucien Hoffmann, and Marco Beyer (2014).
A high-resolution cumulative degree day-based model to simulate phenological development of grapevine.
Am. J. Enol. Vitic., (65:1):72–80.
}
