% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudBurst.R
\name{DD.single.triang}
\alias{DD.single.triang}
\title{Compute the degree-days by the single triangle algorithm}
\usage{
DD.single.triang(t.zero, t.min, t.mean, t.max)
}
\arguments{
\item{t.zero}{threshold temperature for vine growth.}

\item{t.min}{daily minimum air temperature.}

\item{t.max}{daily maximum air temperature.}

\item{x}{xts object containing the input data.}
}
\value{
a vector with the degree-days for vine growth.
}
\description{
Compute the degree-days by the single triangle algorithm
}
