% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudBurst.R
\name{DD.single.triangle.budbreak}
\alias{DD.single.triangle.budbreak}
\title{Cumulative degree days (CDD) by the single triangle algorithm for bud break}
\usage{
DD.single.triangle.budbreak(cdd, start.date)
}
\arguments{
\item{cdd}{cumulative degree days (in Celsius degrees) for vine growth in xts format as provided by
"DD.single.triangle.cumulative" function.}

\item{start.date}{numeric, calculated optimum starting date in day of year.}
}
\value{
the cumulative degree days (in Celsius degrees) for vine growth plus an additional column with
the comulative degree days (in Celsius degrees) for bud break.
}
\description{
Implementation to compute the cumulative degree days by the single triangle algorithm by Nendel (2010)
for bud break.
}
\references{
Nendel, Class (2010). Grapevine bud break prediction for cool winter climates.
Int. J. Biometeorol., 54, 231–241.
}
