% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudBurst.R
\name{DD.single.triangle.cumulative}
\alias{DD.single.triangle.cumulative}
\title{Cumulative degree days (CDD) by the single triangle algorithm}
\usage{
DD.single.triangle.cumulative(data, t.zero, t.min.col, t.mean.col,
  t.max.col)
}
\arguments{
\item{data}{input data in xts format.}

\item{t.zero}{numeric, threshold temperature (in Celsius degrees) for vine growth.}

\item{t.min.col}{numeric, column position in data for the daily minimum air temperature vector in Celsius degrees.}

\item{t.mean.col}{numeric, column position in data for the daily mean air temperature vector in Celsius degrees.}

\item{t.max.col}{numeric, column position in data for the daily maximum air temperature vector in Celsius degrees.}

\item{start.date}{numeric, calculated optimum starting date in day of year.}
}
\value{
list per year for the input data plus an additional column with the cumulative degree days
(in Celsius degrees) for vine growth. The output for each year is a "xts" time series object.
}
\description{
Implementation to compute the cumulative degree days by the single triangle algorithm by Nendel (2010).
}
\references{
Nendel, Class (2010). Grapevine bud break prediction for cool winter climates.
Int. J. Biometeorol., 54, 231–241.
}
