% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux.R
\name{FillNA}
\alias{FillNA}
\title{Fill NA data in time series}
\usage{
FillNA(x)
}
\arguments{
\item{x}{the input time series as xts object}
}
\value{
a time series with the NAs replaced by data according to the na.locf zoo function
}
\description{
Fill NA data in time series
}
