% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{GrowthStage_CDD}
\alias{GrowthStage_CDD}
\title{Cumulative degree days and BBCH growth stages}
\format{
'data.frame':	27 obs. of  3 variables: \cr
$ Growth_stage: int  9 11 12 13 14 15 16 17 18 19 ... \cr
$ Description : chr  "Budburst: green shoot tips clearly visible"  \cr
"First leaf unfolded and spread away from shoot"  \cr
"Two leaves unfolded" "Three leaves unfolded" ... \cr
$ CDD         : num  NA 34.9 58.6 80.8 110.2 ...
}
\usage{
data(GrowthStage_CDD)
}
\description{
A sample dataset containing the cumulative degree days (CDD) for optimized lower,
upper, and heat threshold triplets for Mueller-Thurgau (5°C, 20°C, 22°C),
Riesling  (7°C, 18°C, 24°C) and Pinot noir  (3°C, 20°C, 24°C) per BBCH growth stages
according to Supplementary Table 2 from Molitor and Junk (2019).
}
\references{
Daniel Molitor, Jürgen Junk.
Climate change is implicating a two-fold impact on air temperature increase
in the ripening period under the conditions of the Luxembourgish grapegrowing region.
Vine and Wine Open Access Journal, Oneo One, (3):409-422, 2019.
}
\keyword{data}
