% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux.R
\name{Id.na}
\alias{Id.na}
\title{Find indexes for NA data in time series}
\usage{
Id.na(x)
}
\arguments{
\item{x}{the input time series as xts object.}
}
\value{
a vector with the index for NA data in the time series.
}
\description{
Find indexes for NA data in time series
}
