% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Botrytis.R
\name{lm.botrytis.applied.riesling}
\alias{lm.botrytis.applied.riesling}
\title{Linear regression model applied for predicting the CDD_7;18;24 reaching 5\% disease severity (Botrytis cinerea) for Riesling}
\usage{
lm.botrytis.applied.riesling(
  stat.tmean01,
  stat.tmean02,
  stat.tmean03,
  stat.rain01,
  stat.rain02,
  model
)
}
\arguments{
\item{stat.tmean01}{numeric, vector with two columns: 1) year, and 2) first statistic for mean temperature.}

\item{stat.tmean02}{numeric, vector with two columns: 1) year, and 2) second statistic for mean temperature.}

\item{stat.tmean03}{numeric, vector with two columns: 1) year, and 2) third statistic for mean temperature.}

\item{stat.rain01}{numeric, vector with two columns: 1) year, and 2) first statistic for precipitation.}

\item{stat.rain02}{numeric, vector with two columns: 1) year, and 2) second statistic for precipitation.}

\item{model}{the predefined model e.g. linear regression (lm). Should be a match between the linear
regression coefficients and each statistic in the same definition order.}
}
\value{
data.frame, n obs. of 4 variables (year, predicted CDD for 7, 18 and 24 thresholds reaching 5\% disease severity,
Botrytis cinerea, for Riesling), and lower and upper confidence intervals. n is the total number of years for prediction.
}
\description{
Linear regression model applied for predicting the CDD_7;18;24 reaching 5\% disease severity (Botrytis cinerea) for Riesling
}
