% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{lr.applied.rivaner}
\alias{lr.applied.rivaner}
\title{Linear regression applied}
\usage{
lr.applied.rivaner(
  stat.tmin1,
  stat.tmean1,
  stat.tmax1,
  stat.tmean0,
  stat.rain01,
  stat.rain02,
  stat.tmin0,
  vars,
  coef
)
}
\arguments{
\item{stat.tmin1}{numeric, vector with two columns: 1) year, and 2) statistics for minimun temperature in the previous year}

\item{coef}{}
}
\value{
numeric, a vector of two values: 1) the computation year; and 2) the computed statistic for the
corresponding year.
}
\description{
Linear regression applied
}
