% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{phen.dbbchx.df2}
\alias{phen.dbbchx.df2}
\title{Defining data.frame for yield and disease computation}
\usage{
phen.dbbchx.df2(data.years, year.ini, year.end, phen.luht, bbch.stage)
}
\arguments{
\item{data.years}{list of zoo time series from the input data splitted by years.}

\item{year.ini}{numeric, the initial year for computation.}

\item{year.end}{numeric, the final year for computation.}

\item{phen.luht}{data.frame for the phenology computation by lower, upper and heat thresholds (luht).}

\item{bbch.stage}{numeric, the reference BBCH stage.}
}
\value{
a list with two objects: 1) data.years.subset, the input data.years data.frame subsetted
according to the initial and ending year definition; and 2) phen.d_bbch_x.df2, a data.frame with two
columns per reference BBCH stage defined, the first column is the actual BBCH DOY and the second column is
the DOY for the reference BBCH (a vector of zeros, BBCH DOY reference minus BBCH DOY reference).
}
\description{
Defining data.frame for yield and disease computation
}
