% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenology.R
\name{phenology.stages}
\alias{phenology.stages}
\title{Compute phenological stages}
\usage{
phenology.stages(cdd.phen, ref.data, stage)
}
\arguments{
\item{cdd.phen}{list, cumulative degree days (in Celsius degrees) for vine growth in xts format as
provided by any of the functions "cdd.1Tresh.phenology" or "cdd.2Tresh.phenology" or  "cdd.3Tresh.phenology".}

\item{ref.data}{data.frame, reference dataset to define the phenological stages e.g. "GowthStage_CDD"
dataset.}

\item{stage}{vector, growth stage(s) for which the phenology should be computed. One or more out of
the 27 stages that range from 11 (First leaf unfolded and spread away from shoot) to
89 (Berries ripe for harvest) according to Molitor et al. (2014).}
}
\value{
list per year, with each list containing a data.frame with the phenological stages
for vine growth.
}
\description{
Implementation to compute phenological stages by Molitor et al., (2014).
}
\references{
Daniel Molitor, Jürgen Junk, Danièle Evers, Lucien Hoffmann, and Marco Beyer (2014).
A high-resolution cumulative degree day-based model to simulate phenological development of grapevine.
Am. J. Enol. Vitic., (65:1):72–80.
}
