% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenology.R
\name{Phenology.stages.row}
\alias{Phenology.stages.row}
\title{Compute phenological stages}
\usage{
Phenology.stages.row(cdd.phen, ref.data, stage)
}
\arguments{
\item{cdd.phen}{list, cumulative degree days (in Celsius degrees) for vine growth in xts format as
provided by "DD.single.treshold.phenology" function.}

\item{ref.data}{data.frame, reference dataset to define the phenological stages e.g. "GowthStage_CDD"
dataset.}
}
\value{
list per year, with each list containing a data.frame with the phenological stages
for vine growth.
}
\description{
Implementation to compute phenological stages by Molitor et al., (2014).
}
\references{
Daniel Molitor, Jürgen Junk, Danièle Evers, Lucien Hoffmann, and Marco Beyer (2014).
A high-resolution cumulative degree day-based model to simulate phenological development of grapevine.
Am. J. Enol. Vitic., (65:1):72–80.
}
