% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux.R
\name{plot.na}
\alias{plot.na}
\title{Plot NA data in time series}
\usage{
\method{plot}{na}(x, ids.na)
}
\arguments{
\item{x}{the input time series as xts object}
}
\value{
plots with the NAs highlighted
}
\description{
Plot NA data in time series
}
