% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{plot_na}
\alias{plot_na}
\title{Plot NA data in time series}
\usage{
plot_na(x, ids.na)
}
\arguments{
\item{x}{the input time series as xts object.}

\item{ids.na}{the vector which contains indexes for NA data as provided by the Id.na function.}
}
\value{
plots with the NAs highlighted.
}
\description{
Plot NA data in time series
}
