% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux.R
\name{Raw2xts}
\alias{Raw2xts}
\title{Raw data to xts object}
\usage{
Raw2xts(data)
}
\arguments{
\item{t.min}{daily minimum air temperature}

\item{t.max}{daily maximum air temperature}

\item{t.zero}{threshold temperature for vine growth}
}
\value{
a vector with the degree-days for vine growth
}
\description{
Raw data to xts object
}
