% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{Remich}
\alias{Remich}
\title{Sample time series of meteorological data for Remich station.}
\description{
A sample dataset containing time series of meteorological data for Remich station located in the
Grand-Duchy of Luxembourg obtained by the Institut Viti-vinicole in Remich.
The data covers the period from 1970 to 2017 at daily time interval.
}
\references{
\url{https://agriculture.public.lu/de/weinbau-oenologie.html}
}
\keyword{data}
