% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{window.stat}
\alias{window.stat}
\title{Time window pane analysis}
\usage{
\method{window}{stat}(x, var, fun, t.ini, width, d.ref = FALSE)
}
\arguments{
\item{x}{zoo, time series from the input data for one individual year.}

\item{var}{character, the name of the variable for computing the window pane statistics.}

\item{fun}{character, name of the function for computation in the window pane.}

\item{t.ini}{numeric, the initial number in days from the reference to start the window pane statistics.}

\item{width}{numeric, the width in days of the window pane.}

\item{d68}{numeric or logical, one numeric value for the DOY of the reference BBCH stage to computing the statistic for
the corresponding year or FALSE if reference BBCH stage day is not taken i.e. computation in actual DOY.}
}
\value{
numeric, a vector of two values: 1) the computation year; and 2) the computed statistic for the
corresponding year.
}
\description{
Time window pane analysis
}
