% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yield.R
\name{yield.lr.applied.rivaner}
\alias{yield.lr.applied.rivaner}
\title{Linear regression applied for predicting yield for Rivaner}
\usage{
yield.lr.applied.rivaner(
  stat.tmin1,
  stat.tmean1,
  stat.tmax1,
  stat.tmean0,
  stat.rain01,
  stat.rain02,
  stat.tmin0,
  model
)
}
\arguments{
\item{stat.tmin1}{numeric, vector with two columns: 1) year, and 2) first statistic for min temperature.}

\item{stat.tmean1}{numeric, vector with two columns: 1) year, and 2) first statistic for mean temperature.}

\item{stat.tmax1}{numeric, vector with two columns: 1) year, and 2) first statistic for max temperature.}

\item{stat.tmean0}{numeric, vector with two columns: 1) year, and 2) second statistic for mean temperature.}

\item{stat.rain01}{numeric, vector with two columns: 1) year, and 2) first statistic for precipitation.}

\item{stat.rain02}{numeric, vector with two columns: 1) year, and 2) second statistic for precipitation.}

\item{stat.tmin0}{numeric, vector with two columns: 1) year, and 2) second statistic for min temperature.}

\item{model}{the predefined model e.g. linear regression (lm). Should be a match between the linear
regression coefficients and each statistic in the same definition order.}
}
\value{
data.frame, n obs. of 4 variables (year, predicted yield for Rivaner), and lower and upper values
represent the uncertainty band with 95 percent confidence level. n is the total number of years for prediction.
}
\description{
Linear regression applied for predicting yield for Rivaner
}
