% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{data_water_surface}
\alias{data_water_surface}
\title{A 'sp' object for the water surfaces in Grevenmacher}
\format{Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots \cr
..@ data       :'data.frame':	1373 obs. of  5 variables: \cr
 .. ..$ cat     : int [1:1373] 1 2 3 4 5 6 7 8 9 10 ... \cr
 .. ..$ id      : int [1:1373] 1 1 1 1 1 1 1 1 1 1 ... \cr
 .. ..$ ID_2    : Factor w/ 1353 levels "2?443?282","2?443?944",..: 2 4 9 16 19 35 39 45 51 56 ... \cr
 .. ..$ NATURE  : int [1:1373] 3 3 3 3 3 3 3 3 3 3 ... \cr
 .. ..$ TOPONYME: Factor w/ 4 levels "Bassin d'?puration",..: NA NA NA NA NA NA NA NA NA NA ... \cr
 ..@ polygons   :List of 1373 \cr
 .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots \cr
 .. .. .. ..@ Polygons :List of 1 \cr
 .. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots \cr
 .. .. .. .. .. .. ..@ labpt  : num [1:2] 93462 66333 \cr
 .. .. .. .. .. .. ..@ area   : num 145 \cr
 .. .. .. .. .. .. ..@ hole   : logi FALSE \cr
 .. .. .. .. .. .. ..@ ringDir: int 1 \cr
 .. .. .. .. .. .. ..@ ringDir: int 1 \cr
 .. .. .. .. .. .. ..@ coords : num [1:11, 1:2] 93468 93468 93466 93464 93462 ... \cr
 .. .. .. ..@ plotOrder: int 1 \cr
 .. .. .. ..@ labpt    : num [1:2] 93462 66333 \cr
 .. .. .. ..@ ID       : chr "0" \cr
 .. .. .. ..@ area     : num 145 \cr
 .. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots \cr
 .. .. .. ..@ Polygons :List of 1 \cr
 .. .. .. ..[list output truncated]}
\source{
\url{  https://data.public.lu/en/datasets/bd-l-tc-surface-deau/#_}
}
\usage{
data(data_water_surface)
}
\description{
A dataset containing a 'sp' object for the water surfaces in the District of Grevenmacher
in the Grand-Duchy of Luxembourg. The dataset contains streches from the Moselle (Mosel) and the S{\^u}re rivers.
The data is provided by the Luxembourgish Data Platform, and provided as a SpatialPolygonsDataFrame from "sp" package.
}
\details{
The spatial representation for the data corresponds to a SpatialPolygonsDataFrame with
coordinate reference system (CRS) EPSG:2169 - Luxembourg 1930 / Gauss - Projected.\cr

The dataset identifies natural or artificial water surface, permanent or not. The field
NATURE is defined as:  0 = stream surface; 1 = stagnant water surface; 2 = wet zone; 3 = basin;
4 = purification basin; 5 = pool; 6 = fish farming. \cr
The selection criteria for defining the NATURE field is: \cr
0: surface of stream of permanent flow. Minimum width 3.5 m. \cr
1: permanent water surface without flow: lake, pond, pond whose smallest dimension is greater than 20 m.
Dams are treated in this class and are updated if their area has increased 50\%. \cr
2: wetland area greater than 2 Ha, marsh, aquatic vegetation. \cr
3: open basin. Treatment plant basins, swimming pools or fish farms are treated with
a particular attribute value. Most small dimension must be greater than 10 m and the outline masonry. \cr
4: sewage treatment plant basin, the smallest dimension of which must be greater than 10 m and the
masonry outline. \cr
5: swimming pool whose smallest dimension must be greater than 10 m. \cr
6: fish pond with the smallest dimension greater than 10 m and the masonry outline.
}
\examples{
data(data_water_surface)
str(data_water_surface, list.len = 15)

#' plot stfdf object
library(sp)
spplot(data_water_surface, zcol="id", scales = list(draw = TRUE),
       colorkey = FALSE)

}
\keyword{data}
