% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudBurst.R
\name{DD.single.triangle}
\alias{DD.single.triangle}
\title{Degree days by the single triangle algorithm for bud break}
\usage{
DD.single.triangle(t.zero, t.min, t.mean, t.max)
}
\arguments{
\item{t.zero}{numeric, threshold temperature (in Celsius degrees) for vine growth.}

\item{t.min}{daily minimum air temperature vector in Celsius degrees.}

\item{t.max}{daily maximum air temperature vector in Celsius degrees.}
}
\value{
a vector with the degree-days (in Celsius degrees) for vine growth.
}
\description{
Implementation to compute the degree days by the single triangle algorithm by Nendel (2010) for bud break.
}
\references{
Nendel, Class (2010). Grapevine bud break prediction for cool winter climates. Int. J. Biometeorol., 54, 231–241.
}
