% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{GrowthStage_CDD}
\alias{GrowthStage_CDD}
\title{Cumulative degree days and growth stages}
\format{'data.frame':	27 obs. of  3 variables: \cr
$ Growth_stage: int  9 11 12 13 14 15 16 17 18 19 ... \cr
$ Description : chr  "Budburst: green shoot tips clearly visible"  \cr
"First leaf unfolded and spread away from shoot"  \cr
"Two leaves unfolded" "Three leaves unfolded" ... \cr
$ CDD         : num  NA 34.9 58.6 80.8 110.2 ...}
\usage{
data(GrowthStage_CDD)
}
\description{
A sample dataset containing the cumulative degree days (CDD) with optimized lower,
upper, and heat threshold triplets (5, 20, 22°C) per growth stages
according to Table 4 from Molitor et al. (2014).
}
\references{
Daniel Molitor, Jürgen Junk, Danièle Evers, Lucien Hoffmann, and Marco Beyer.
A high-resolution cumulative degree day-based model to simulate phenological development of grapevine.
Am. J. Enol. Vitic., (65:1):72–80, 2014.
}
\keyword{data}
