% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{id.na}
\alias{id.na}
\title{Find indexes for NA data in time series}
\usage{
id.na(x)
}
\arguments{
\item{x}{the input time series as xts object.}
}
\value{
a vector with the index for NA data in the time series.
}
\description{
Find indexes for NA data in time series
}
