% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aux.R
\name{Raw2xts}
\alias{Raw2xts}
\title{Raw data to xts object}
\usage{
Raw2xts(data)
}
\arguments{
\item{data}{the dataframe to convert to xts time series.}
}
\value{
the xts object for the input dataframe.
}
\description{
Raw data to xts object
}
